Abstract PheWASTable into separate component
