+++ src/pages/VariantPage.js
@@ -5,9 +5,10 @@ import { Query } from 'react-apollo';
 import gql from 'graphql-tag';
 
 import { PheWAS } from 'ot-charts';
-import { PageTitle, Heading, SubHeading, OtTable } from 'ot-ui';
+import { PageTitle, Heading, SubHeading } from 'ot-ui';
 
 import BasePage from './BasePage';
+import PheWASTable from '../tables/PheWASTable';
 
 function hasAssociations(data) {
   return (
@@ -31,33 +32,6 @@ const pheWASQuery = gql`
     }
   }
 `;
-
-const tableColumns = [
-  {
-    label: 'nCases',
-    key: 'nCases',
-  },
-  {
-    label: 'nTotal',
-    key: 'nTotal',
-  },
-  {
-    label: 'pval',
-    key: 'pval',
-  },
-  {
-    label: 'studyId',
-    key: 'studyId',
-  },
-  {
-    label: 'traitCode',
-    key: 'traitCode',
-  },
-  {
-    label: 'traitReported',
-    key: 'traitReported',
-  },
-];
 
 const VariantPage = ({ match }) => (
   <BasePage>
@@ -118,7 +92,7 @@ const VariantPage = ({ match }) => (
         return hasAssociations(data) ? (
           <Fragment>
             <PheWAS data={data} />
-            <OtTable columns={tableColumns} data={data.pheWAS.associations} />
+            <PheWASTable data={data} />
           </Fragment>
         ) : null;
       }}
